+++ CHANGELOG.md
@@ -13,3 +13,11 @@ All notable autonomous changes should be recorded here.
 ## 2026-07-25T15:48:32.923456+00:00
 
 - Added a simple Python utility to generate a hardware inventory template and added a test.
+
+## 2026-08-01T14:00:00.000000+00:00
+
+- Added a simple Python utility to generate a FreeCAD case design checklist and added a test.
+
+## 2026-08-05T10:00:00.000000+00:00
+
+- Added a simple Python utility to generate an ESP32/e-paper pinout helper and added a test.